srfge

+++ hw3p3/hw3p3/Checksum.h
@@ -13,7 +13,7 @@ public:
 	DWORD crc_table[256];
 
 	Checksum();
-	DWORD CRC32(unsigned char* buf, size_t len);
+	DWORD CRC32(unsigned char* buf, UINT64 len);
 
 
 };
